Изменение принимаемого типа данных при умножении и делении

+++ Vector.cs
@@ -28,12 +28,12 @@
             return new Vector(-a.X, -a.Y);
         }
 
-        public static Vector operator *(Vector a, int b)
+        public static Vector operator *(Vector a, float b)
         {
             return new Vector(a.X * b, a.Y * b);
         }
 
-        public static Vector operator /(Vector a, int b)
+        public static Vector operator /(Vector a, float b)
         {
             return new Vector(a.X / b, a.Y / b);
         }
